requirements.txt
streamlit
tensorflow-cpu
pandas
numpy
scikit-learn
joblib